Issue #7: move Flask app building from __init__.py to views.py
