+++ openeo_driver/__init__.py
@@ -1,6 +1,1 @@
-from flask import Flask
-app = Flask(__name__)
-app.config['APPLICATION_ROOT'] = '/openeo'
-
-from ._version import __version__
-import openeo_driver.views
+from  openeo_driver._version import __version__
