+++ openeo_driver/server.py
@@ -1,7 +1,7 @@
 import gunicorn.app.base
 from gunicorn.six import iteritems
 
-from openeo_driver import app
+from openeo_driver.views import app
 
 """
 Script to start a production server. This script can serve as the entry-point for doing spark-submit.
